+++ devicemanagement/CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [SW-18590] Added firmware update functional and stress tests.
 - [SW-15748] Handling DM events in userspace
   - Adding DM event processing thread for each device
+- [SW-18613] Added support to change the PMIC FW metadata in the flash image and re-generate it for testing.
 ### Changed
 - [SW-18398b] Update DevErrorEvent implementation for the skipList usage to add/remove EventType entry from the default skipList without requiring to replace the complete skipList.
 [SW-18797] Improvements for event processor